pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "groq-cli"
version = "0.1.0"
dependencies = ["groq", "python-dotenv"]

[project.scripts]
groq = "groq_cli:main"
